Move the DbTransformation classes to a new package
