+++ src/main/java/uk/gov/gchq/magmacore/demo/ExampleSignsRdl.java
@@ -18,8 +18,8 @@ import java.util.List;
 
 import uk.gov.gchq.hqdm.rdf.iri.HQDM;
 import uk.gov.gchq.hqdm.rdf.iri.RDFS;
-import uk.gov.gchq.magmacore.service.DbChangeSet;
-import uk.gov.gchq.magmacore.service.DbCreateOperation;
+import uk.gov.gchq.magmacore.service.transformation.DbChangeSet;
+import uk.gov.gchq.magmacore.service.transformation.DbCreateOperation;
 
 /**
  * Functions for creating systems using MagmaCore and HQDM.
