+++ src/main/java/uk/gov/gchq/magmacore/service/transformation/DbTransformation.java
@@ -12,13 +12,15 @@
  * the License.
  */
 
-package uk.gov.gchq.magmacore.service;
+package uk.gov.gchq.magmacore.service.transformation;
 
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+
+import uk.gov.gchq.magmacore.service.MagmaCoreService;
 
 /**
  * Class representing an invertible ordered sequence of change sets.
